+++ agent/temboardagent/cli/routes.py
@@ -1,5 +1,8 @@
 import logging
 import sys
+from itertools import chain
+
+from bottle import default_app
 
 from ..routing import get_routes
 from ..toolkit.app import SubCommand
@@ -14,13 +17,9 @@ class Routes(SubCommand):
     """ Show HTTP routing table."""
 
     def main(self, args):
-        routes = get_routes()
-        for route in sorted(routes, key=route_sort_key):
-            method = route['http_method']
-            path = route['path'].decode('utf-8')
-            mod = sys.modules[route['module']]
-            fn = getattr(mod, route['function'])
-            doc = fn.__doc__ or ''
+        routes = chain(iter_routes(get_routes()), iter_bottle_routes())
+        for path, method, callback in sorted(routes):
+            doc = callback.__doc__ or ''
             try:
                 title = doc.splitlines()[0]
             except IndexError:
@@ -30,5 +29,14 @@ class Routes(SubCommand):
         return 0
 
 
-def route_sort_key(route):
-    return route['path'], route['http_method']
+def iter_routes(routes):
+    for route in routes:
+        mod = sys.modules[route['module']]
+        callback = getattr(mod, route['function'])
+        yield route['path'].decode('ascii'), route['http_method'], callback
+
+
+def iter_bottle_routes():
+    bottle = default_app()
+    for route in bottle.routes:
+        yield route.rule, route.method, route.callback