+++ agent/temboardagent/web/__init__.py
@@ -1,0 +1,3 @@
+from .service import HTTPDService
+
+__all__ = ['HTTPDService']